+++ src/main/kotlin/io/kotest/assertions/arrow/core/Nel.kt
@@ -1,10 +1,50 @@
-package io.kotest.assertions.arrow.nel
+package io.kotest.assertions.arrow.core
 
 import arrow.core.NonEmptyList
+import io.kotest.inspectors.forAll
+import io.kotest.inspectors.forAny
+import io.kotest.inspectors.forAtLeast
+import io.kotest.inspectors.forAtLeastOne
+import io.kotest.inspectors.forAtMost
+import io.kotest.inspectors.forAtMostOne
+import io.kotest.inspectors.forExactly
+import io.kotest.inspectors.forNone
+import io.kotest.inspectors.forOne
+import io.kotest.inspectors.forSome
 import io.kotest.matchers.Matcher
 import io.kotest.matchers.MatcherResult
 import io.kotest.matchers.should
 import io.kotest.matchers.shouldNot
+
+fun <A> NonEmptyList<A>.forAll(fn: (A) -> Unit): Unit =
+  all.forAll(fn)
+
+fun <A> NonEmptyList<A>.forOne(fn: (A) -> Unit): Unit =
+  all.forOne(fn)
+
+fun <A> NonEmptyList<A>.forExactly(k: Int, fn: (A) -> Unit): Unit =
+  all.forExactly(k, fn)
+
+fun <A> NonEmptyList<A>.forSome(fn: (A) -> Unit): Unit =
+  all.forSome(fn)
+
+fun <A> NonEmptyList<A>.forAny(f: (A) -> Unit): Unit =
+  all.forAny(f)
+
+fun <A> NonEmptyList<A>.forAtLeastOne(f: (A) -> Unit): Unit =
+  all.forAtLeastOne(f)
+
+fun <A> NonEmptyList<A>.forAtLeast(k: Int, fn: (A) -> Unit): Unit =
+  all.forAtLeast(k, fn)
+
+fun <A> NonEmptyList<A>.forAtMostOne(f: (A) -> Unit): Unit =
+  all.forAtMostOne(f)
+
+fun <A> NonEmptyList<A>.forAtMost(k: Int, fn: (A) -> Unit): Unit =
+  all.forAtMost(k, fn)
+
+fun <T> NonEmptyList<T>.forNone(f: (T) -> Unit): Unit =
+  all.forNone(f)
 
 fun <A> NonEmptyList<A>.shouldContainOnlyNulls(): Unit =
   should(containOnlyNulls())
